added repository pattern

+++ app/frontend/Models/ViewModels/CartViewModel.cs
@@ -7,12 +7,13 @@
         public long ISBN { get; set; }
         public string GenreName { get; set; }
         public string TypeName { get; set; }
-        public double Prices { get; set; }
+        public decimal Prices { get; set; }
         public string Url { get; set; }
 
         public string CartItem_Id { get; set; }
 
         public int quantity { get; set; }
         public long PriceId { get; set; }
+
     }
 }
